requirements.txt
# Core dependencies
llama-index==0.10.42
llama-index-core==0.10.42
llama-index-llms-openai==0.1.22
llama-index-embeddings-huggingface==0.2.1
llama-index-vector-stores-qdrant==0.2.8
llama-index-readers-file==0.1.23

# OpenAI and embeddings
openai==1.30.5
sentence-transformers==2.7.0

# Vector database
qdrant-client==1.9.1

# Web interface
gradio==4.36.0

# Configuration and utilities
python-dotenv==1.0.1
pydantic==2.7.2

# Document processing
pypdf==4.2.0
python-multipart==0.0.9

# HTTP and networking
requests==2.32.3
httpx==0.27.0

# Data processing
numpy==1.26.4
pandas==2.2.2

# Utilities
tqdm==4.66.4
rich==13.7.1
